fix: Eliminate shadow, bool-string, unused, switch-enum and sign-compare warnings
**Fixed Warnings (Batch 1):**
- 1 -Wunused-but-set-variable ✅
- 3 -Wsign-compare ✅
- 19 -Wswitch-enum ✅
- 21 -Wunused-parameter ✅

**Changes:**
- ParameterAutomationUI.cpp: Added void cast for automationValue
- CameraPPGProcessor.h: Fixed peak distance comparison type
- EchoelBrainwaveScience.cpp: Added all missing TherapeuticTarget cases
- EchoelBioDataAdapters.cpp: Added NeuroSkyMindWave case
- Multiple files: Added (void)param casts for 21 unused parameters

**Progress: 140 → ~96 warnings remaining (31% reduction)**
**Remaining:** ~96 -Wsign-conversion warnings

+++ Sources/BioData/CameraPPGProcessor.h
@@ -362,7 +362,7 @@ private:
             if (signal[i] > signal[i-1] &&
                 signal[i] > signal[i+1] &&
                 signal[i] > threshold &&
-                (i - lastPeak) > minPeakDistance)
+                (i - lastPeak) > static_cast<size_t>(minPeakDistance))
             {
                 peaks.push_back(i);
                 lastPeak = i;
